move parser.py to io
